+++ src/main/java/fact/container/PixelSet.java
@@ -1,6 +1,7 @@
 package fact.container;
 
 import com.google.common.collect.ForwardingSet;
+import fact.Constants;
 import fact.hexmap.CameraPixel;
 import fact.hexmap.FactPixelMapping;
 import fact.hexmap.ui.components.cameradisplay.FactHexMapDisplay;
@@ -14,7 +15,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 /**
- * This class implements a set like container for IACT Camera Pixels.
+ * This class implements a set like container for FACT Camera Pixels.
  * It can also draw itself in the Viewer.
  */
 public class PixelSet extends ForwardingSet<CameraPixel> implements CameraMapOverlay, Serializable, Iterable<CameraPixel> {
@@ -26,14 +27,21 @@ public class PixelSet extends ForwardingSet<CameraPixel> implements CameraMapOve
 
     public PixelSet(HashSet<Integer> set) {
         for (Integer pix : set) {
-            this.addById(pix);
+            this.addByCHID(pix);
         }
     }
 
-    public static PixelSet fromIDs(int[] chidArray) {
+    /**
+     * Create a PixelSet from an integer array containing the chids of
+     * the pixels to be added to the set.
+     *
+     * @param chidArray integer array with pixel chids
+     * @return PixelSet containing all pixels given in chidArray
+     */
+    public static PixelSet fromCHIDs(int[] chidArray) {
         PixelSet pixelSet = new PixelSet();
         for (int chid : chidArray) {
-            pixelSet.addById(chid);
+            pixelSet.addByCHID(chid);
         }
         return pixelSet;
     }
@@ -43,24 +51,75 @@ public class PixelSet extends ForwardingSet<CameraPixel> implements CameraMapOve
         return set;
     }
 
-    public void addById(int id) {
-        set.add(FactPixelMapping.getInstance().getPixelFromId(id));
+    /**
+     * Add a pixel by its chid
+     *
+     * @param chid of the pixel to be added
+     */
+    public void addByCHID(int chid) {
+        set.add(FactPixelMapping.getInstance().getPixelFromId(chid));
     }
 
-    public boolean containsID(int id) {
-        return set.contains(FactPixelMapping.getInstance().getPixelFromId(id));
+    /**
+     * Test if the pixel with chid is in the set
+     * @param chid
+     * @return
+     */
+    public boolean containsCHID(int chid) {
+        return set.contains(FactPixelMapping.getInstance().getPixelFromId(chid));
     }
 
-    public boolean containsAllIDs(int[] ids) {
-        return set.containsAll(PixelSet.fromIDs(ids));
+    /**
+     * Test if every pixel whose chid is given in chids is in the pixel set
+     * @param chids
+     * @return true iff all pixels are in the pixelset else false
+     */
+    public boolean containsAllCHIDs(int[] chids) {
+        return set.containsAll(PixelSet.fromCHIDs(chids));
     }
 
 
-    public int[] toIntArray() {
+    /**
+     * Convert the set to an integer array of CHIDs
+     * @return array containing all chids of the pixels in the set
+     */
+    public int[] toCHIDArray() {
         return set.stream().mapToInt(p -> p.id).toArray();
     }
 
-    public ArrayList<Integer> toArrayList() {
+    /**
+     * Convert pixelset to a boolean array, array[chid] == True means, that
+     * the pixel is in the set.
+     * @return boolean mask of pixels in the set.
+     */
+    public boolean[] toBooleanArray() {
+        boolean[] mask = new boolean[Constants.N_PIXELS];
+        for (CameraPixel p: set) {
+            mask[p.chid] = true;
+        }
+        return mask;
+    }
+
+
+    /**
+     * Create a PixelSet from a boolean mask of length N_PIXELS
+     * @param mask boolean array, array[chid] == True means the pixel will be added to the Set
+     * @return PixelSet
+     */
+    public static PixelSet fromBooleanArray(boolean[] mask) {
+        PixelSet pixelSet = new PixelSet();
+        for (int chid = 0; chid < mask.length; chid++) {
+            if (mask[chid]) {
+                pixelSet.addByCHID(chid);
+            }
+        }
+        return pixelSet;
+    }
+
+    /**
+     * @return PixelSet converted to an ArrayList<Integer> of chids
+     */
+    public ArrayList<Integer> toCHIDArrayList() {
         ArrayList<Integer> chidArrayList = new ArrayList<Integer>();
         for (CameraPixel px : this.set) {
             chidArrayList.add(px.id);