+++ src/main/java/fact/datacorrection/InterpolateTimeSeries.java
@@ -36,6 +36,12 @@ public class InterpolateTimeSeries implements Processor {
     @Parameter(required = false, description = "The minimum number of neighboring pixels required for interpolation", defaultValue = "3")
     public int minPixelToInterpolate = 3;
 
+    @Parameter(required = false, description = "The key for the resulting badPixelSet.")
+    public String badPixelKey = "badPixels";
+
+    @Parameter(required = false, description = "The key to the timestamp of the Event.")
+    public String timeStampKey = "timestamp";
+
     private FactPixelMapping pixelMap = FactPixelMapping.getInstance();
 
 
@@ -44,17 +50,7 @@ public class InterpolateTimeSeries implements Processor {
         Utils.isKeyValid(item, dataKey, double[].class);
         double[] data = (double[]) item.get(dataKey);
 
-        ZonedDateTime timeStamp = null;
-
-        if (item.containsKey("UnixTimeUTC") == true) {
-            Utils.isKeyValid(item, "UnixTimeUTC", int[].class);
-            int[] eventTime = (int[]) item.get("UnixTimeUTC");
-            timeStamp = Utils.unixTimeUTCToZonedDateTime(eventTime);
-        } else {
-            // MC Files don't have a UnixTimeUTC in the data item. Here the timestamp is hardcoded to 1.1.2000
-            // => The 12 bad pixels we have from the beginning on are used.
-            timeStamp = ZonedDateTime.of(2000, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC);
-        }
+        ZonedDateTime timeStamp = Utils.getTimeStamp(item, timeStampKey);
 
         PixelSet badPixelSet = calibService.getBadPixels(timeStamp);
 
@@ -67,7 +63,7 @@ public class InterpolateTimeSeries implements Processor {
         }
 
         item.put(dataOutputKey, data);
-        item.put("badPixels", badPixelSet);
+        item.put(badPixelKey, badPixelSet);
         return item;
     }