+++ src/main/java/fact/io/hdureader/OptionalTypesMap.java
@@ -132,7 +132,7 @@ public class OptionalTypesMap<K, V> extends ForwardingMap<K, V> {
     }
 
     /**
-     * Get the value for the keyword in the header as long if it exists and the value can be cast to Long.
+     * Get the value for the given key as long if it exists and the value can be cast to Long.
      *
      * @return an optional holding the value for the key
      */
@@ -145,13 +145,39 @@ public class OptionalTypesMap<K, V> extends ForwardingMap<K, V> {
     }
 
     /**
-     * Get the value for the keyword in the header as long array if it exists and the value can be cast to long[].
+     * Get the value for the given key as long array if it exists and the value can be cast to long[].
      *
      * @return an optional holding the value for the key
      */
     public Optional<long[]> getLongArray(K key) {
         try {
             return Optional.ofNullable((long[]) delegate().get(key));
+        } catch (ClassCastException e) {
+            return Optional.empty();
+        }
+    }
+
+    /**
+     * Get the value for the given key as boolean  if it exists and the value can be cast to boolean.
+     *
+     * @return an optional holding the value for the key
+     */
+    public Optional<Boolean> getBoolean(K key) {
+        try {
+            return Optional.ofNullable((Boolean) delegate().get(key));
+        } catch (ClassCastException e) {
+            return Optional.empty();
+        }
+    }
+
+    /**
+     * Get the value for the given key as boolean array if it exists and the value can be cast to boolean[].
+     *
+     * @return an optional holding the value for the key
+     */
+    public Optional<boolean[]> getBooleanArray(K key) {
+        try {
+            return Optional.ofNullable((boolean[]) delegate().get(key));
         } catch (ClassCastException e) {
             return Optional.empty();
         }