+++ src/test/java/fact/utils/RemappingTest.java
@@ -5,6 +5,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertArrayEquals;
 
 /**
  * Created by kaibrugge on 12.08.14.
@@ -42,5 +43,15 @@ public class RemappingTest {
         assertTrue("Remapping went wrong. Element should be 890", remappedData[393 * roi] == 0);
     }
 
+    @Test
+    public void testReRemapping() {
+        Remapping re = new Remapping();
+        int roi = data.length / 1440;
+        assertTrue("First array element should contain 0", data[0] == 0);
+        re.remapFromSoftIdToChid(data, remappedData);
+        short[] reremappedData = new short[1440 * 300];
+        re.remapFromChidToSoftId(remappedData, reremappedData);
 
+        assertArrayEquals(data, reremappedData);
+    }
 }